# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Card Scanner Pro API
servers:
  - url: /api/v1
    description: Canonical versioned base path.
  - url: /api
    description: >-
      Legacy base path, retained for backward compatibility and deprecated in favor of
      /api/v1. Responses on this prefix include RFC 8594 Deprecation and Link
      (rel="successor-version") headers.
...
paths:
...
  /search/contacts:
    post:
      operationId: searchContacts
      tags: [search]
      summary: Advanced multi-condition contact search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchContactsInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchContactsResult"
  /saved-searches:
    get:
      operationId: listSavedSearches
      tags: [search]
      summary: List the caller's saved filters and views
      parameters:
        - name: entityType
          in: query
          required: false
          schema:
            type: string
        - name: kind
          in: query
          required: false
          schema:
            type: string
            enum: [filter, view]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedSearchList"
    post:
      operationId: createSavedSearch
      tags: [search]
      summary: Save a filter or view
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedSearchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedSearch"
  /saved-searches/{id}:
    patch:
      operationId: updateSavedSearch
      tags: [search]
      summary: Rename or update a saved search
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedSearchUpdate"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedSearch"
    delete:
      operationId: deleteSavedSearch
      tags: [search]
      summary: Delete a saved search
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /recent-searches:
    get:
      operationId: listRecentSearches
      tags: [search]
      summary: List the caller's recent searches
      parameters:
        - name: entityType
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentSearchList"
    delete:
      operationId: clearRecentSearches
      tags: [search]
      summary: Clear the caller's recent searches
      parameters:
        - name: entityType
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Contact:
      type: object
      required: [id, companyId, status, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        fullName:
          type: ["string", "null"]
        arabicName:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        contactCompany:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        mobile:
          type: ["string", "null"]
        officePhone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        gpsAccuracy:
          type: ["number", "null"]
        duplicateOfId:
          type: ["integer", "null"]
        linkedin:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [new, contacted, quotation_sent, negotiation, won, lost, qualified, interested, proposal_sent, archived]
        leadScore:
          type: ["integer", "null"]
        leadTemperature:
          type: ["string", "null"]
          enum: [hot, warm, cold, null]
        aiReasoning:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        seniority:
          type: ["string", "null"]
        enrichmentSummary:
          type: ["string", "null"]
        talkingPoints:
          type: array
          items:
            type: string
        enrichedAt:
          type: ["string", "null"]
          format: date-time
        followUpDate:
          type: ["string", "null"]
          format: date
        followUpTime:
          type: ["string", "null"]
        cardImageUrl:
          type: ["string", "null"]
        eventId:
          type: ["integer", "null"]
        eventName:
          type: ["string", "null"]
        assignedToId:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    SearchCondition:
      type: object
      required: [field, operator]
      properties:
        field:
          type: string
          enum: [name, company, email, phone, industry, country, tags, notes, event, employee]
        operator:
          type: string
          enum: [contains, notContains, equals, notEquals, startsWith, endsWith, isEmpty, isNotEmpty]
        value:
          type: ["string", "number", "null"]
    SearchContactsInput:
      type: object
      required: [conditions]
      properties:
        combinator:
          type: string
          enum: [AND, OR]
          default: AND
        conditions:
          type: array
          items:
            $ref: "#/components/schemas/SearchCondition"
        sort:
          type: ["string", "null"]
        page:
          type: integer
        limit:
          type: integer
    SearchContactsResult:
      type: object
      required: [contacts, total, page, limit]
      properties:
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    SavedSearch:
      type: object
      required: [id, entityType, kind, name, payload, createdAt, updatedAt]
      properties:
        id:
          type: integer
        entityType:
          type: string
        kind:
          type: string
          enum: [filter, view]
        name:
          type: string
        payload: {}
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SavedSearchList:
      type: object
      required: [savedSearches]
      properties:
        savedSearches:
          type: array
          items:
            $ref: "#/components/schemas/SavedSearch"
    SavedSearchInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        kind:
          type: string
          enum: [filter, view]
          default: filter
        entityType:
          type: string
          default: contacts
        payload: {}
    SavedSearchUpdate:
      type: object
      properties:
        name:
          type: string
        payload: {}
    RecentSearch:
      type: object
      required: [id, entityType, label, payload, createdAt]
      properties:
        id:
          type: integer
        entityType:
          type: string
        label:
          type: string
        payload: {}
        createdAt:
          type: string
          format: date-time
    RecentSearchList:
      type: object
      required: [recentSearches]
      properties:
        recentSearches:
          type: array
          items:
            $ref: "#/components/schemas/RecentSearch"
...