# lib/api-zod/src/generated/types/recentSearch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Card Scanner Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface RecentSearch {
  id: number;
  entityType: string;
  label: string;
  payload: unknown;
  createdAt: Date;
}