# lib/api-zod/src/generated/types/savedSearchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Card Scanner Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { SavedSearchInputKind } from './savedSearchInputKind';

export interface SavedSearchInput {
  name: string;
  kind?: SavedSearchInputKind;
  entityType?: string;
  payload?: unknown;
}