# lib/api-zod/src/generated/types/savedSearchUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Card Scanner Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface SavedSearchUpdate {
  name?: string;
  payload?: unknown;
}