# lib/api-zod/src/generated/types/searchCondition.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Card Scanner Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { SearchConditionField } from './searchConditionField';
import type { SearchConditionOperator } from './searchConditionOperator';

export interface SearchCondition {
  field: SearchConditionField;
  operator: SearchConditionOperator;
  /** @nullable */
  value?: string | number | null;
}