# lib/api-zod/src/generated/types/searchContactsResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Card Scanner Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { Contact } from './contact';

export interface SearchContactsResult {
  contacts: Contact[];
  total: number;
  page: number;
  limit: number;
}